Restructure client and parser modules
