+++ llm_utils/pipelines/lora_autotag.py
@@ -1,0 +1,6 @@
+"""Pipeline utilities for generating captions via LoRA models."""
+
+
+def generate_captions(*args, **kwargs):
+    """Placeholder function for caption generation."""
+    raise NotImplementedError("lora_autotag pipeline not implemented")
