+++ pyproject.toml
@@ -56,4 +56,4 @@ run-bert = "llm_utils.inference.run_bert:main"
 train-t5 = "llm_utils.training.train_t5:main"
 run-t5 = "llm_utils.inference.run_t5:main"
 prepare-dataset = "llm_utils.training.prepare_dataset:main"
-llm-request = "llm_utils.interfacing.llm_request:main"
+llm-request = "llm_utils.clients.openai_client:main"
